requirements.txt
black==22.3.0
flake8==4.0.1
isort==5.10.1
pre-commit==2.19.0
pytest>=7.0.0
pyyaml>=6.0
